css/cssom-view/client-props-zoom.html

<!DOCTYPE html>
<title>getBoundingClientRect for elements with css zoom</title>
<link rel="author" title="Yotam Hacohen" href="mailto:yotha@chromium.org">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://drafts.csswg.org/css-viewport/">"
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
  <head>
    <style>
      div {
        width: 64px;
        height: 64px;
        background-color: blue
      }
      div.a {
        zoom: 4.0;
      }
    </style>
  </head>
  <body>
    <div id="no_zoom"></div>
    <div class="a" id="with_zoom"></div>
    <script>
      setup(() => {
        window.noZoom = document.getElementById("no_zoom");
        window.withZoom = document.getElementById("with_zoom");
      });
      test(function() {
        assert_true(!!noZoom, "no zoom target exists");
        assert_true(!!withZoom, "zoom target exists");
      });
      test(function() {
        assert_equals(noZoom.clientTop, withZoom.clientTop, 'clientTop');
        assert_equals(noZoom.clientLeft, withZoom.clientLeft, 'clientLeft');
        assert_equals(noZoom.clientWidth, withZoom.clientWidth, 'clientWidth');
        assert_equals(noZoom.clientHeight, withZoom.clientHeight, 'clientHeight');
      });
    </script>
  </body>
